Rename modules to client + server instead of request + serve + socket
