+++ crates/lune-std-net/src/lib.rs
@@ -4,9 +4,8 @@ use mlua::prelude::*;
 
 use lune_utils::TableBuilder;
 
-mod request;
-mod serve;
-mod socket;
+mod client;
+mod server;
 mod url;
 
 const TYPEDEFS: &str = include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/types.d.luau"));
